requirements.txt
streamlit>=1.31,<1.37
snowflake-connector-python>=3.10.0
pandas>=2.1
python-dotenv>=1.0